- option to manage logrotate
- disable manage logrotate for Debian by default

+++ manifests/init.pp
@@ -20,6 +20,7 @@ class freeradius (
   Boolean $preserve_mods                                       = true,
   Boolean $correct_escapes                                     = true,
   Boolean $manage_logpath                                      = true,
+  Boolean $manage_logrotate                                    = true,
   Optional[String] $package_ensure                             = 'installed',
   String $radacctdir                                           = $freeradius::params::radacctdir,
 ) inherits freeradius::params {
@@ -428,37 +429,39 @@ class freeradius (
     }
   }
 
-  logrotate::rule { 'radacct':
-    path          => "${freeradius::fr_logpath}/radacct/*/*.log",
-    rotate_every  => 'day',
-    rotate        => 7,
-    create        => false,
-    missingok     => true,
-    compress      => true,
-    postrotate    => "kill -HUP `cat ${freeradius::fr_pidfile}`",
-    sharedscripts => true,
-  }
+  if $manage_logpath {
+    logrotate::rule { 'radacct':
+      path          => "${freeradius::fr_logpath}/radacct/*/*.log",
+      rotate_every  => 'day',
+      rotate        => 7,
+      create        => false,
+      missingok     => true,
+      compress      => true,
+      postrotate    => "kill -HUP `cat ${freeradius::fr_pidfile}`",
+      sharedscripts => true,
+    }
 
-  logrotate::rule { 'checkrad':
-    path          => "${freeradius::fr_logpath}/checkrad.log",
-    rotate_every  => 'week',
-    rotate        => 1,
-    create        => true,
-    missingok     => true,
-    compress      => true,
-    postrotate    => "kill -HUP `cat ${freeradius::fr_pidfile}`",
-    sharedscripts => true,
-  }
+    logrotate::rule { 'checkrad':
+      path          => "${freeradius::fr_logpath}/checkrad.log",
+      rotate_every  => 'week',
+      rotate        => 1,
+      create        => true,
+      missingok     => true,
+      compress      => true,
+      postrotate    => "kill -HUP `cat ${freeradius::fr_pidfile}`",
+      sharedscripts => true,
+    }
 
-  logrotate::rule { 'radiusd':
-    path          => "${freeradius::fr_logpath}/radius*.log",
-    rotate_every  => 'week',
-    rotate        => 26,
-    create        => true,
-    missingok     => true,
-    compress      => true,
-    postrotate    => "kill -HUP `cat ${freeradius::fr_pidfile}`",
-    sharedscripts => true,
+    logrotate::rule { 'radiusd':
+      path          => "${freeradius::fr_logpath}/radius*.log",
+      rotate_every  => 'week',
+      rotate        => 26,
+      create        => true,
+      missingok     => true,
+      compress      => true,
+      postrotate    => "kill -HUP `cat ${freeradius::fr_pidfile}`",
+      sharedscripts => true,
+    }
   }
 
   # Placeholder resource for dh and random as they are dynamically generated, so they
